feat(adminLandingPage): fixed components hierarchy and unnasigned card
